Deluge defaults update

+++ rootfs/opt/init.sh
@@ -19,15 +19,19 @@ adduser --system --disabled-password --no-create-home --uid "$PUID" xyz
 addgroup --system --gid "$PGID" xyz
 adduser xyz xyz
 
-# Handling default configs
+Handling default configs
 echo "Handling default config for Deluge..."
 if [[ ! -f $DELUGE_CONFIG/core.conf ]]; then
-	cp /defaults/core.conf $DELUGE_CONFIG/core.conf
+	cp /defaults/*.conf $DELUGE_CONFIG/
 fi
 echo "Installing the FileBot plugin if it's missing..."
 if [[ ! -f $DELUGE_CONFIG/plugins/FileBotTool-*$PYTHON_VER.egg ]]; then
 	echo "FileBot plugin is missing... Installing..."
-	rm $DELUGE_CONFIG/plugins/FileBotTool-*.egg
+	if [[ ! -f $DELUGE_CONFIG/plugins/FileBotTool-*.egg ]]; then
+		mkdir -p $DELUGE_CONFIG/plugins
+	else
+		rm $DELUGE_CONFIG/plugins/FileBotTool-*.egg
+	fi
 	cp /defaults/plugins/FileBotTool-*.egg $DELUGE_CONFIG/plugins/$PLUGIN_NAME$PYTHON_VER.egg
 fi
